backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
torch   # sentence-transformers will require a backend; pip may install automatically but include it
pdfplumber
python-docx